#37287 #37353 (2.2.0.89) BIT_FLAGS型を定義し、型の置換を継続中。 / Define BIT_FLAGS, ongoing type replacement.

+++ src/bldg.c
@@ -2190,7 +2190,7 @@ static void get_questinfo(IDX questnum, bool do_init)
 	p_ptr->inside_quest = old_quest;
 
 	/* Print the quest info */
-	sprintf(tmp_str, _("クエスト情報 (危険度: %d 階相当)", "Quest Information (Danger level: %d)"), quest[questnum].level);
+	sprintf(tmp_str, _("クエスト情報 (危険度: %d 階相当)", "Quest Information (Danger level: %d)"), (int)quest[questnum].level);
 
 	prt(tmp_str, 5, 0);
